requirements.txt
git+https://github.com/Rapptz/discord-ext-menus
aiofiles==0.6.0
aiohttp==3.7.4
aiosqlite==0.17.0
akinator.py==5.0.0
beautifulsoup4==4.9.3
discord.py==1.7.3
googletrans==3.1.0a0
humanize==3.10.0
jishaku==2.2.0
lavalink==3.1.4
matplotlib==3.4.1
numpy==1.20.2
Pillow==8.3.2
prsaw==0.3.1
PyNaCl==1.4.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.25.1
urllib3==1.26.5
wavelink==0.9.9
wikipedia==1.4.0
wolframalpha==5.0.0
youtube-dl==2021.6.6
